backend/requirements.txt
Django==5.0
djangorestframework
django-cors-headers
psycopg2-binary==2.9.9
requests
PyJWT
djangorestframework-simplejwt
social-auth-app-django
# Add other dependencies as needed
